+++ api/index.js
@@ -1,10 +1,9 @@
 /* eslint-disable no-console */
 import express from 'express';
 import winston from 'winston';
-import signin from './routes/signIn';
-import signup from './routes/signUp';
-import articles from './routes/articles';
-import viewFeeds from './routes/viewFeed';
+import userRoute from './routes/auth';
+import articleRoute from './routes/articles';
+import commentRoute from './routes/commentArticle';
 import auth from './middleware/auth';
 
 const app = express();
@@ -17,10 +16,9 @@ process.on('unhandledRejection', (ex) => {
 
 winston.add(new winston.transports.File({ filename: 'logfile.log' }));
 
-app.use('/api/v1/auth/signup', signup);
-app.use('/api/v1/auth/signin', signin);
-app.use('/api/v1/feeds', auth, viewFeeds);
-app.use('/api/v1/articles', auth, articles);
+app.use('/api/v1/auth/', userRoute);
+app.use('/api/v1/', auth, articleRoute);
+app.use('/api/v1/', auth, commentRoute);
 
 require('../api/helpers/prod')(app);
