+++ api/models/article.js
@@ -5,6 +5,13 @@ const articles = [
     title: 'Frontend',
     article: 'Do you want to be a front end developer',
     authorId: 1,
+    comments: [
+      {
+        comment_id: 1,
+        authorId: 1,
+        comment: 'good article',
+      },
+    ],
   },
   {
     id: 2,
@@ -12,6 +19,13 @@ const articles = [
     title: 'Backend',
     article: 'Do you want to be a Backend developer',
     authorId: 2,
+    comments: [
+      {
+        comment_id: 2,
+        authorId: 1,
+        comment: 'good article',
+      },
+    ],
   },
   {
     id: 3,
@@ -19,6 +33,13 @@ const articles = [
     title: 'Backend',
     article: 'Do you want to be a Backend developer',
     authorId: 3,
+    comments: [
+      {
+        comment_id: 3,
+        authorId: 1,
+        comment: 'good article',
+      },
+    ],
   },
 ];
