--- api/routes/signUp.js
@@ -1,8 +1,0 @@
-import express from 'express';
-import signup from '../controllers/signup';
-
-const router = express();
-
-router.post('/', signup);
-
-export default router;
