--- api/routes/viewFeed.js
@@ -1,9 +1,0 @@
-import express from 'express';
-import viewArticles from '../controllers/viewArticles';
-
-const router = express();
-
-router.get('/', viewArticles);
-
-
-export default router;
